refactored structure to add sensors
