+++ custom_components/starling_home_hub/__init__.py
@@ -3,6 +3,7 @@
 For more details about this integration, please refer to
 https://github.com/ThomasLomas/ha-starlinghomehub
 """
+
 from __future__ import annotations
 
 from homeassistant.config_entries import ConfigEntry
@@ -14,7 +15,7 @@ from .api import StarlingHomeHubApiClient
 from .const import DOMAIN, PLATFORMS
 from .coordinator import StarlingHomeHubDataUpdateCoordinator
 
-# https://developers.home-assistant.io/docs/config_entries_index/#setting-up-an-entry
+
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up this integration using UI."""
