+++ custom_components/starling_home_hub/const.py
@@ -1,7 +1,11 @@
-"""Constants for starling_home_hub."""
+"""All constants used globally across the integration."""
+
 from logging import Logger, getLogger
 
 from homeassistant.const import Platform
+
+from custom_components.starling_home_hub.integrations.sensor import SENSOR_PLATFORMS
+from custom_components.starling_home_hub.integrations.smoke_detector import SMOKE_DETECTOR_PLATFORMS
 
 LOGGER: Logger = getLogger(__package__)
 
@@ -16,3 +20,8 @@ PLATFORMS: list[Platform] = [
     Platform.CLIMATE,
     # Platform.CAMERA,
 ]
+
+DEVICE_TYPES_TO_PLATFORMS = {
+    "smoke_detector": SMOKE_DETECTOR_PLATFORMS,
+    "sensor": SENSOR_PLATFORMS,
+}
