+++ custom_components/starling_home_hub/models/api/device/base.py
@@ -1,4 +1,5 @@
 """This contains the base device class."""
+
 from dataclasses import dataclass
