+++ custom_components/starling_home_hub/models/api/devices.py
@@ -1,4 +1,5 @@
-"""This module contains the Devices class."""
+"""This module contains the Devices response class."""
+
 from dataclasses import dataclass
 
 from custom_components.starling_home_hub.models.api.device.base import BaseDevice
