+++ custom_components/starling_home_hub/models/coordinator.py
@@ -1,4 +1,5 @@
 """This module contains the CoordinatorData class."""
+
 from dataclasses import dataclass
 
 from custom_components.starling_home_hub.models.api.device import Device
